Feat(migration): relocate models
**What**

- relocated models from server/app.py to db/models
- change path for migration
- change structure app.py
- add index.py blueprint
- fix test
- add config

related to  #10
