+++ backend/db/models/admin_info.py
@@ -1,0 +1,9 @@
+from db.models import db
+
+
+class AdminInfo(db.Model):
+    __tablename__ = 'AdminInfo'
+    id = db.Column(db.Integer, primary_key=True)
+    email = db.Column(db.String(50), unique=True, nullable=False)
+    psw = db.Column(db.String(500), nullable=False)
+
