+++ backend/db/models/booking_info.py
@@ -1,0 +1,10 @@
+from db.models import db
+
+
+class BookingInfo(db.Model):
+    __tablename__ = 'BookingInfo'
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(50), nullable=False)
+    email = db.Column(db.String(500), nullable=False)
+    topic = db.Column(db.String(500))
+    slots_inf = db.relationship('Slots', backref='info', lazy='dynamic')
