+++ backend/server/__init__.py
@@ -1,0 +1,6 @@
+from server.app import create_app
+
+app = create_app()
+
+if __name__ == '__main__':
+    app.run(debug=True)
