--- backend/server/hello_json.py
@@ -1,3 +1,0 @@
-msg_hello = {
-    "greeting": "Hello World"
-}
